accessible/tests/mochitest/relations/test_ui_modalprompt.html

<html>

<head>
  <title>Modal prompts</title>
  <link rel="stylesheet" type="text/css"
        href="chrome://mochikit/content/tests/SimpleTest/test.css" />

  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"></script>

  <script type="application/javascript"
          src="../common.js"></script>
  <script type="application/javascript"
          src="../relations.js"></script>
  <script type="application/javascript"
          src="../role.js"></script>
  <script type="application/javascript"
          src="../events.js"></script>
  <script type="application/javascript"
          src="../browser.js"></script>

  <script type="application/javascript">
    function hasTabModalPrompts() {
      try {
        return SpecialPowers.getBoolPref("prompts.tab_modal.enabled");
      } catch (ex) {
        return false;
      }
    }

    function showAlert() {
      this.eventSeq = [
        {
          type: EVENT_SHOW,
          match: function(aEvent) {
            return aEvent.accessible.role == ROLE_DIALOG;
          }
        }
      ];

      this.invoke = function showAlert_invoke() {
        window.setTimeout(
          function() {
            currentTabDocument().defaultView.alert("hello");
          }, 0);
      };

      this.check = function showAlert_finalCheck(aEvent) {
        var dialog = aEvent.accessible.DOMNode;
        var info = dialog.ui.infoBody;
        testRelation(info, RELATION_DESCRIPTION_FOR, dialog);
        testRelation(dialog, RELATION_DESCRIBED_BY, info);
      };

      this.getID = function showAlert_getID() {
        return "show alert";
      };
    }

    // gA11yEventDumpToConsole = true; // debug

    var gQueue = null;
    function doTests() {
      gQueue = new eventQueue();
      gQueue.push(new showAlert());
      gQueue.onFinish = function() {
        synthesizeKey("VK_RETURN", {}, browserWindow());
        closeBrowserWindow();
      };
      gQueue.invoke(); // will call SimpleTest.finish()
    }

    if (!hasTabModalPrompts()) {
      todo(false, "Test disabled when tab modal prompts are not enabled.");
    } else {
      SimpleTest.waitForExplicitFinish();
      openBrowserWindow(doTests);
    }
  </script>

</head>

<body id="body">

  <a target="_blank"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=661293"
     title="The tabmodalprompt dialog's prompt label doesn't get the text properly associated for accessibility">
    Mozilla Bug 661293
  </a>
  <br>
  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test">
  </pre>

</body>
</html>
